feat(craft-app): move init command
