+++ tests/commands/test_init.py
@@ -28,7 +28,7 @@ import pytest
 from craft_cli import CraftError
 from flake8.api.legacy import get_style_guide
 
-from charmcraft.commands.init import DEFAULT_PROFILE, PROFILES, InitCommand
+from charmcraft.application.commands.init import DEFAULT_PROFILE, PROFILES, InitCommand
 from charmcraft.models.charmcraft import Project
 from charmcraft.utils import S_IXALL
 from tests.test_infra import get_python_filepaths
